refactor: flatten wos/research/ and wos/plan/ into top-level package
