--- wos/plan/__init__.py
@@ -1,1 +1,0 @@
-"""WOS plan skill support modules."""
