--- wos/research/__init__.py
@@ -1,1 +1,0 @@
-"""WOS research skill support modules."""
